engine: split the shared forwarder out and rename the app router
Disambiguate the engine modules and untangle the router-to-router dependency:

- New server/engine_forward.py owns the shared proxy layer (_proxy/_forward, the
  per-child keep-alive pool, cancel-on-hangup, the heal-retry and the sentinels).
  Both engine routers import it, so neither reaches into the other.
- routers/engine.py -> routers/app_engine.py: the /api/engine (warm app worker)
  surface, no longer told apart from the /api/engines template surface by a
  single 's'. engines.py keeps only the template routes.

Pure move + rename; no behavior change. URLs are unchanged.
